Added loading state for currency on offer-overview
