+++ client/src/plus/offer-overview/offer-hero/index.tsx
@@ -2,8 +2,35 @@ import { CRUD_MODE } from "../../../constants";
 import Mandala from "../../../ui/molecules/mandala";
 import "./index.scss";
 
-function OfferHero({ plusAvailable }: { plusAvailable: boolean }) {
+function OfferHero({
+  currency,
+  plusAvailable,
+}: {
+  currency: string;
+  plusAvailable: boolean;
+}) {
   const animate = !CRUD_MODE;
+  let container;
+
+  //No currency, Pricing info still loading.Display spinner
+  if (!currency) {
+    container = (
+      <>
+        <div className="cash-container">
+          <div className="cash-spinner">
+            $<br />€<br />£<br />$<br />€<br />£<br />$<br />€
+          </div>
+        </div>
+        <span>5.00</span>
+      </>
+    );
+  } else {
+    container = new Intl.NumberFormat("en-US", {
+      style: "currency",
+      currency: currency,
+    }).format(5);
+  }
+
   return (
     <div className="dark offer-hero">
       <header className="container offer-hero-header">
@@ -14,7 +41,8 @@ function OfferHero({ plusAvailable }: { plusAvailable: boolean }) {
           {(plusAvailable && (
             <>
               <h2>
-                Support MDN <u>and</u> make it your own. For just $5 a month.
+                Support MDN <u>and</u> make it your own. For just {container} a
+                month.
               </h2>
               <div className="button-wrapper">
                 <a href="#subscribe" className="button-primary">
